requirements.txt
python-telegram-bot==22.6
python-dotenv==1.0.1
openpyxl==3.1.5
Pillow==11.0.0
imageio==2.36.0
imageio-ffmpeg==0.5.1
